feat(core): optimize component loading with dynamic imports (#328)
### TL;DR
Implemented dynamic imports for components and restructured the notification system for better performance.

### What changed?
- Added dynamic imports using `next/dynamic` for various components to enable code splitting
- Restructured notification components into separate files for better organization
- Removed Vercel Analytics and React Query DevTools
- Optimized component rendering by conditionally rendering the import employees dialog
- Split notification-bell component into smaller, more manageable components

### How to test?
1. Verify notifications still appear and function when clicking the bell icon
2. Check that employee imports still work correctly
3. Confirm account requests and account creation dialogs function as expected
4. Ensure toasts and confirmation dialogs appear properly
5. Verify the loading performance improvements in the network tab

### Why make this change?
This change improves the application's initial load time and overall performance by implementing code splitting through dynamic imports. The notification system restructuring provides better maintainability and separation of concerns, while removing unused analytics and dev tools reduces the bundle size.
