+++ src/app/(dashboard)/(home)/accounts/add-account-button.tsx
@@ -1,5 +1,4 @@
 'use client'
-import AddAccountForm from '@/app/(dashboard)/(home)/accounts/add-account-form'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
@@ -10,7 +9,13 @@ import {
   DialogTrigger,
 } from '@/components/ui/dialog'
 import { Plus } from 'lucide-react'
+import dynamic from 'next/dynamic'
 import { useState } from 'react'
+
+const AddAccountForm = dynamic(
+  () => import('@/app/(dashboard)/(home)/accounts/add-account-form'),
+  { ssr: false },
+)
 
 const AddAccountButton = () => {
   const [isOpen, setIsOpen] = useState(false)
