+++ src/app/(dashboard)/(home)/accounts/request/account-request.tsx
@@ -1,5 +1,4 @@
 'use client'
-import AccountRequestList from '@/app/(dashboard)/(home)/accounts/request/account-request-list'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
@@ -12,7 +11,14 @@ import {
 import getPendingAccounts from '@/queries/get-pending-accounts'
 import { createBrowserClient } from '@/utils/supabase'
 import { useQuery } from '@supabase-cache-helpers/postgrest-react-query'
+import dynamic from 'next/dynamic'
 import { useState } from 'react'
+
+const AccountRequestList = dynamic(
+  () =>
+    import('@/app/(dashboard)/(home)/accounts/request/account-request-list'),
+  { ssr: false },
+)
 
 const AccountRequest = () => {
   const [isOpen, setIsOpen] = useState<boolean>(false)
