+++ src/app/layout.tsx
@@ -1,12 +1,23 @@
 import { GeistSans } from 'geist/font/sans'
 import ThemeProvider from '@/providers/ThemeProvider'
 import NextTopLoader from 'nextjs-toploader'
-import { Analytics } from '@vercel/analytics/react'
 import './globals.css'
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import ReactQueryProvider from '@/providers/ReactQueryProvider'
-import { Toaster } from '@/components/ui/toaster'
-import ConfirmationDialog from '@/components/confirmation-dialog/confirmation-dialog'
+import dynamic from 'next/dynamic'
+
+const Toaster = dynamic(
+  () => import('@/components/ui/toaster').then((mod) => mod.Toaster),
+  {
+    ssr: false,
+  },
+)
+
+const ConfirmationDialog = dynamic(
+  () => import('@/components/confirmation-dialog/confirmation-dialog'),
+  {
+    ssr: false,
+  },
+)
 
 const defaultUrl = process.env.VERCEL_PROJECT_PRODUCTION_URL
   ? `https://${process.env.VERCEL_PROJECT_PRODUCTION_URL}`
@@ -45,10 +56,8 @@ export default function RootLayout({
               {children}
               <Toaster />
               <ConfirmationDialog />
-              <Analytics />
-              {/* ^^ remove this if you are not deploying to vercel. See more at https://vercel.com/docs/analytics  */}
             </main>
-            <ReactQueryDevtools initialIsOpen={false} />
+            {/* <ReactQueryDevtools initialIsOpen={false} /> */}
           </ReactQueryProvider>
         </ThemeProvider>
       </body>
